+++ src/Application/Models/StatusAplicacaoPagamentoPix.cs
@@ -1,0 +1,9 @@
+namespace Application.Models;
+
+public enum StatusAplicacaoPagamentoPix
+{
+    Aplicado = 0,
+    JaAplicado = 1,
+    SemResultadoConclusivo = 2,
+    RequerReconciliacao = 3
+}